+++ packages/app/src/components/AppBar/CourseSearchBar.vue
