+++ packages/app/src/components/Popup/OverwriteLockedSectionPopup.vue
@@ -1,5 +1,5 @@
 <template>
-    <v-dialog v-model="getNoTimetablePopup" persistent max-width="290">
+    <v-dialog v-model="getOverwriteLockedSectionPopup" persistent max-width="290">
         <v-card>
             <v-card-title
                 class="white--text align-end cardTitle"
@@ -12,18 +12,17 @@
             </v-card-title>
             <v-card-text class="text--primary">
                 <div class="textStyle">
-                    We are unable to generate a pair of timetables based on the
-                    changes you made. We will roll back your previous timetable.
+                    You are trying to overwrite a section that is currently locked,
+                    do you want to unlock it and proceed?
                 </div>
             </v-card-text>
             <v-card-actions>
                 <v-spacer></v-spacer>
-                <v-btn
-                    color="blue darken-1"
-                    text
-                    @click="setNoTimetablePopup(false)"
-                >
-                    I understand
+                <v-btn color="blue darken-1" text @click="onClickProceed">
+                    Proceed
+                </v-btn>
+                <v-btn color="blue darken-1" text @click="onClickCancel">
+                    Cancel
                 </v-btn>
             </v-card-actions>
         </v-card>
@@ -34,16 +33,19 @@
 import { mapGetters, mapMutations } from "vuex";
 
 export default {
-    data() {
-        return {
-            logo: require("../assets/barbackground.png"),
-        };
-    },
     computed: {
-        ...mapGetters(["getNoTimetablePopup"]),
+        ...mapGetters(["getOverwriteLockedSectionPopup"]),
     },
     methods: {
-        ...mapMutations(["setNoTimetablePopup"]),
+        ...mapMutations(["setOverwriteLockedSectionPopup"]),
+        onClickProceed() {
+            this.$emit("Proceed");
+            this.setOverwriteLockedSectionPopup(false);
+        },
+        onClickCancel() {
+            this.$emit("Cancel");
+            this.setOverwriteLockedSectionPopup(false);
+        },
     },
 };
 </script>
@@ -55,7 +57,4 @@ export default {
 .textStyle {
     margin-top: 20px;
 }
-/* .buttonStyle{
-  min-width:0px;
-} */
 </style>
