+++ packages/app/src/components/Timetable/TimetableCourseCard.vue
@@ -60,10 +60,9 @@
 
 <script>
 import { mapGetters, mapActions } from "vuex";
-import CourseSectionPicker from "../components/CourseSectionPicker";
+import CourseSectionPicker from "../Popup/CourseSectionPicker";
 
 export default {
-  name: "timetable-course-card",
   components: {
     CourseSectionPicker
   },
