+++ packages/app/src/components/Timetable/WeekdaySwitch.vue
